kernel: rename device.h -> blk_device.h since it only contains block device declarations
